[User Education] Simplify User Activity precondition, allow zero timeout
This removes mouse clicks and touch events from the list of events that
can cause an IPH to be delayed (now it is only typing) since that was
likely too aggressive, and also ensures that when the "minimum time from
user activity" parameter is zero, the precondition always succeeds.

This will be used to run several study arms varying or removing the time
between typing and when a heavyweight promo can show.

Bug: 369886363
Change-Id: Ie79c3e918703b9ca61ceb3611f0278c4235303cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6382617
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1436958}