Chrome Incorrectly Selects Fallback Font When src:local("")
When the src: local(..) inside @font-face is specified as empty
(@font-face { src: local(""); }), Chrome incorrectly selects a fallback
font. Ideally, it should fallback on the standard font. The issue arises
from Chrome's font matching method, which allows it to find a font with
a blank src path, leading to the fallback on an incorrect font. Other
browsers, like Safari and Firefox, render the default font in case of an
empty src

Bug: 342067852
Change-Id: Id2ffbcaf63812bef2b8f71b90e64cc05b8f79855
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6344210
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Gaurav Kumar <gauravkum@microsoft.com>
Reviewed-by: Vinay Singh <vinaysingh@microsoft.com>
Reviewed-by: Dileep Maurya <dileepmaurya@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1436963}