viz: Use the pass backing color space as the draw layer color space
This change makes `SkiaRenderer::CurrentDrawLayerColorSpace` return the
actual color space of the current render pass backing, rather than the
compositing color space.

We see a mismatch in the scanout render pass backing case where
`DirectRenderer::CalculateRenderPassRequirements` uses the output color
space, but `SkiaRenderer::CurrentDrawLayerColorSpace` always uses the
compositing color space.

In the case where we have:
- HDR quad resources
- but no quads that require blending,
on Windows we end up with:
- no HDR color conversion layer
- a render pass backing with linear scRGB color space
- and `SkiaRenderer::CurrentDrawLayerColorSpace` returning the render
  pass compositing color space (extended sRGB).

When we draw the TextureDrawQuad with the HDR resource, we end up
drawing it directly to the render pass backing with a color conversion
filter going from the resource color space to the pass compositing
color space, but the backing is actually scRGB. This results in the
quad appearing darker than expected.

Bug: 397995970
Change-Id: Ia9824927510a9e7b9a0fe2e48e3da797b7486686
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299975
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: Sida Zhu <zhusida@bytedance.com>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1436973}