Addresses a layout regression in the Privacy Sandbox Dialogs
The embedded privacy policy iframe was causing unexpected scrolling and
content overflow. The issue stemmed from the iframe's default
positioning behavior, which allowed it to overflow its container and
affect the overall layout of the dialog even when it was supposed to be
"invisible." This resulted in an inconsistent UX when a user would
expand and collapse the "learn more" section.

This CL explicitly positions the iframe using top: 0 and left: 0,
ensuring it stays at the top-left corner of the container and doesn't
interfere with other elements. We also set the iframe's position to
fixed so that we don't have to worry about the parent of the iframe
somehow affecting its positioning later.

CL also adds test coverage for the previous regression bug:
crbug.com/388420268

Bug: 390002434
Change-Id: I2f2eec679098c186b0894319cb52121f00a4e1ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6305383
Reviewed-by: Kevin Graney <kmg@google.com>
Commit-Queue: Jacob Stanley <jacobstanley@google.com>
Reviewed-by: Abe Boujane <boujane@google.com>
Reviewed-by: Christine Ying <chrstne@google.com>
Auto-Submit: Jacob Stanley <jacobstanley@google.com>
Cr-Commit-Position: refs/heads/main@{#1436975}