Remove the recursion in PageDiscardingHelper::DiscardMultiplePages
The algorithm is changed to always pop candidates from the back of
the vector and to keep trying until a discard attempt is successful
or the vector is empty.

Removing the recursion will also avoid uncessessary recalculations
of the candidates vector and the `page_node_footprint_kb` map.

Bug: 404584647
Change-Id: I4febc830e339ccf15ddc443650a95bcebe9923fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6378067
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1436994}