Fix/prevent race condition on profile re-loading
This CL switches to using `CreateTaskRunnerForResource()` when the BTM
service is using a persistent/on-disk database, avoiding a scenario
where a new (different) task runner is used to open the BTM db while the
original task runner is in the process of closing it.

`CreateTaskRunnerForResource()` prevents this by only creating a new
task runner if there isn't already one associated with the db file. If
there is already a task runner associated with the db file, it returns a
reference to that one.

Reusing the same task runner means the task to open the db won't start
until the task to close it has completed.

Bug: 369635654
Change-Id: I780ceced23d01c3e1441f0ad8c1802ade4a59c0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6386752
Reviewed-by: Ryan Tarpine <rtarpine@chromium.org>
Commit-Queue: Joshua Hood <jdh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437050}