[fedcm] Preview and fill verified email suggestions
In addition to resolving the conditional get request promise, we also release to the website the text value in the form field upon selection.

Before that, on hovering of a selection, we also preview its value in the website, which doesn't actually release the value to the website.

From this, we'll work on generalizing the filling to cover other use cases (such as passwords) as well as filling more than one field at a time (e.g. emails and names within a single selection).

Bug: 380367784
Change-Id: Icdcc7c7a6b21eb944d0f62a9271fb97f527509eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6377384
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Sam Goto <goto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437062}