Reopen the glic widget after user logs in from a paused state.
This CL creates a new class, LogInAndOpen, which holds two states for the user's log in status - kLogIn and kPostLogIn. If the user account is paused, show the Reauth page and set the status of log_in_and_open_ to kLogIn until the user logs back in. After the user logs in, reopen the widget regardless if there's an attached browser or not.

However, this approach leads to one tradeoff, which can happen in the following user flow: The user clicks the button, we show the Reauth page, but the user decides not to sign in, and instead reauths somewhere else with a different access token other than kGlic - this still reopens the widget.
Will follow up with a fix for this in crbug.com/405230722.

Bug: 399730016, 405230722
Change-Id: I7bfac57f821dce828e860d433e633e1210299b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6377427
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Commit-Queue: Sana Akbani <sanaakbani@google.com>
Cr-Commit-Position: refs/heads/main@{#1437082}