[ntp][ogb] Add dark mode request url query param
As part of loading the OGB, the new `dark` styling parameter is now
included in the request.

In the OGB's script, when the async OGB parts feature is enabled, we now
wait for an event to instruct the OGB is ready before posting the loaded
event. As a result, the NTP element will dispatch appearance update
requests only once we know the OGB component is ready.

Bug: 373569279
Change-Id: Icadcae8db86c78f62cd60fb563e8ea1eb214276b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6283085
Reviewed-by: Riley Tatum <rtatum@google.com>
Commit-Queue: Roman Arora <romanarora@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437087}