[PWA] Remove ExternallyManagedAppInstallTask dependencies [2/N]
This CL is the 2nd step in the chain to remove any dependency on
ExternallyManagedAppInstallTask. This is done on production by updating
the struct that stored an instance of the task to only store an instance
of the ExternalInstallOption and the corresponding callback. When an
installation needs to happen, the WebAppCommandScheduler is used to
invoke an ExternalAppResolutionCommand directly with the callback, thus
completely bypassing the requirement of an intermediary task like state.

This CL also updates existing documentation that mentions the
ExternallyManagedAppInstallTask to instead refer to the command that is
called.

A future CL will remove all traces of the task code altogether.

Change-Id: Iaf88ae4c9c709a0a1ee72c63163f0f4bdd3dd313
Bug: 328431086
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6357998
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437102}