[crd host] Pass disconnect error details and location via mojo
This allows the desktop process on Windows to report the error details
and location to the website client.

`error_details` is straightforward since it's just a string, but the
location is tricky, since `base::Location` does not have a backing store
for the function name and file name -- it just references to string
constants in the data segment, which means it's impossible to
reconstruct a `base::Location` when it gets passed across the process
boundary. To make this work, this CL introduces a `SourceLocation`
class, which is similar to `base::Location`, but has an optional backing
store, which is allocated when it is reconstructed with the mojo struct.

Bug: 382334458
Change-Id: I0285eaca34c44bc16053c565114c393510fa5512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6358815
Reviewed-by: Joe Downing <joedow@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Auto-Submit: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437150}