Create DraggingTabsSession to handle dragging tabs within a tabstrip.
DraggingTabsSession handles the work of the kDraggingTabs state for
TabDragController. This functionally boils down to implementing
MoveAttached. TabDragController creates a new DraggingTabsHelper each
time it enters kDraggingTabs (or, in one case, when it enters
kWaitingToDragTabs instead), and destroys it when exiting that state.

This removes quite a few responsibilities from TabDragController - most
of the detailed business logic bits are in this domain, even though it's
a minority of the code size. It's also more testable, since it doesn't
have to deal with creating windows or running move loops. A compact
interface with minimal dependencies, hiding lots of requirements? It's
unit test heaven. Followup CLs will set up a test harness and convert
TabDragControllerInteractiveUITests into DraggingTabsSessionUnitTests.

Bug: 382754501
Change-Id: I79ed701e8a961901ccc19ffe140a08657d105e2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6345457
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437187}