[omnibox][search aggregator] Add size parameter to image URLs for googleusercontent
Image requests for googleusercontent expect a size parameter in the URL.
If the size parameter is missing, the rendered image may have unexpected
behavior including inverted colors. This change checks for the URL in
the image_url and appends a size parameter if needed.

Bug: 406000460
Change-Id: I93037bb1ecf89cde63664e4d0f161e6e110fd222
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387024
Commit-Queue: Alex Chen <alexwchen@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437202}