[ReferenceTarget] Ensure .list and .form reflect the shadow host
Per the resolution of https://github.com/WICG/webcomponents/issues/1072,
we need to ensure that the .list and .form properties return the host
element instead of null when they refer to an element in a shadow via
reference target.

For the .list property, we update HTMLInputElement::listForBinding to
return the host if the referenced element is a valid data list element.

For the .form property, we introduce ListedElement::FormForBinding to
return the host if the referenced element is a valid form element. The
related .idl files are updated to bind formForBinding (which ultimately
calls ListedElement::FormForBinding) instead of formOwner, which now
returns the actual associated form element (either in the shadow dom via
reference target, or in the light dom via ID reference)

Note: 1. While we change the return type of the .form property from
HTMLFormElement to HTMLElement, we return HTMLElement only when the flag
ShadowRootReferenceTarget is enabled and a reference target is
specified, so no web pages would be affected until they start using
reference targets.

2. While this PR addresses the association from the referencing element
to the referenced form, i.e., a reset button can reset the associated
form in the shadow dom (as verified in
third_party/blink/web_tests/external/wpt/shadow-dom/reference-target/tentative/form.html),
the association from the referenced form to the referencing element is
still not working, eg. the referenced form's .elements property doesn't
include the referencing element. The issue will be addressed in a
separate CL. More details can be found in
https://issues.chromium.org/issues/401039567#comment2

Bug: 399819896, 401039567
Change-Id: I6ca3bed9b39c7eaf4bd10f514cad89246b628b3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6349470
Reviewed-by: Dan Clark <daniec@microsoft.com>
Commit-Queue: Alex Keng <shihken@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437221}