[E2E] Allow creating EdgeToEdgeLayoutCoordinator from Context
Change the ctor signature from Activity to context. Activity is only
used for Activity#getWindow within #shouldPadDisplayCutout, so we could
do a casting within the static method.

When context is not an Activity, how window insets is handle should be
irrelevant. For code correctness, when the given context is not an
activity, we'll always disable padding for the display cutout.

Bug: 378704318
Change-Id: I06aa05badd401ce87afc432e9dcfc84025159fc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6383841
Reviewed-by: Charles Hager <clhager@google.com>
Auto-Submit: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1437227}