Add UMA to know how likely DoInitStream() blocks for Google Search
This is similar to https://crrev.com/c/5785518, but specifically targets
Google Search.

This metric will help us understand the P90 latency between FetchStart
and RequestStart, as described in go/no-yield-analysis (internal
document).

Instead of directly recording a histogram in HttpNetworkTransaction
[^1], this CL plumbs timing information (stream initialization delay)
as part of LoadTimingInternalInfo, and then records the histogram
in GWSPageLoadMetricsObserver.

The purpose of using GWSPageLoadMetricsObserver is to ensure that
metrics are recorded under the exact same conditions as other metrics
which use GWSPageLoadMetricsObserver. This ensures consistency and
allows for better comparisons down the road.

LoadTimingInternalInfo is a new structure I'm introducing to group
internal load timing information. We typically put timing information in
LoadTimingInfo, but that should be reserved for web-exposed data.

The key difference between LoadTimingInfo and LoadTimingInternalInfo is
that LoadTimingInternalInfo is only provided if the request has
`trusted_params` (that is, the loader is trusted).

LoadTimingInternalInfo will be used to group other internal load timing
data that we want to measure in the future.

[^1]: https://crrev.com/c/6332172

Bug: 359404121,393980912

Change-Id: If9feabae35c14b416602040ca24b6d6cc66f9e68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6335162
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437230}