[Instant message] Use last focused window for tab group removed message
For tab group removed messages, it's hard to find the window that the
tab group originally belonged to. It's even harder now given that we
are queuing the messages for ~1 second for aggregation purposes.
Using the last focused window gives a better approximation here to
show the instant message.

Bug: 400794347
Change-Id: I4a1a84eaffbbdefbd7f1091ea08883eda5fa11ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6382053
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437231}