extensions: Make drag and drop .crx install work on desktop android
We take the file info on the drop event on android, and the drag event
on other platforms, because otherwise we can see an empty object. This
is due to fundamental platform restrictions at least on android.

To test it:

1. Patch this change and its dependencies.
2. Enable developer mode
3. Drag and drop a .crx file to chrome://extensions
4. Reload

The extension is shown!

Bug: 392777363
Change-Id: I198650aef494bbca353a08018abf635233456b3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6370848
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437252}