[UPMLoginDbSunset] Record metrics for the old GMS no passwords dialog
This dialog is shown when a user tries to access the password manager,
but they are on a device with a version of Google Play Services that
is too old for full UPM support. It is also only shown for users who
have either downloaded or didn't have passwords left over in Chrome.

The histogram records the user decision to accept to update or not
in the dialog.

Bug: 405988411
Change-Id: I537ae9428ff0db0a09b8db9bb85dc0f49f3cc237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387916
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1437381}