[Extensions] Install bubble: remove empty space if no how-to-use text
One overlooked aspect of MV2 to MV3 migration is that the how-to-use
text from the extension installed bubble does not handle the "action"
action type from the manifest, so MV3 extensions don't have that text.

The dialog allocates space for it but no text is shown there, resulting
in a weird empty space.

This CL is a temporary fix to remove that space if the text is empty
until the right text is decided...

Screenshot: https://imgur.com/a/dtj7Nuw

Bug: 405148986
Change-Id: I02508e27fd0231d00986b92c116fecd0ab53b6aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6384458
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1437389}